+++ include/AssemblerParser.hpp
@@ -34,6 +34,7 @@ public:
     friend class AssemblerOptimizer;
     friend class AssemblerSimulatedOps;
     friend class AssemblerGenerator;
+    friend std::vector<uint8_t> emitO45(AssemblerParser&, const std::string&);
     AssemblerParser(const std::vector<AssemblerToken>& tokens);
     AssemblerParser(const std::vector<AssemblerToken>& tokens, const std::map<std::string, uint32_t>& predefinedSymbols);
     void pass1();